Move culumn definition tests to the correct folder
Tests that work with actual models need to access the database and must
as such be under integration test folder.
